--- app/src/main/java/com/sunasterisk/boringweather/data/MainRepository.kt
@@ -1,4 +1,0 @@
-package com.sunasterisk.boringweather.data
-
-interface MainRepository {
-}
