+++ app/src/main/java/com/sunasterisk/boringweather/data/model/City.kt
@@ -1,4 +1,4 @@
-package com.sunasterisk.boringweather.data.local.model
+package com.sunasterisk.boringweather.data.model
 
 import android.content.res.Resources
 import androidx.annotation.RawRes
