+++ app/src/main/java/com/sunasterisk/boringweather/data/model/CurrentWeather.kt
@@ -1,4 +1,4 @@
-package com.sunasterisk.boringweather.data.local.model
+package com.sunasterisk.boringweather.data.model
 
 data class CurrentWeather(
     val currentWeather: HourlyWeather = HourlyWeather(),
