+++ app/src/main/java/com/sunasterisk/boringweather/data/repository/CurrentRepository.kt
@@ -1,8 +1,8 @@
-package com.sunasterisk.boringweather.data
+package com.sunasterisk.boringweather.data.repository
 
 import com.sunasterisk.boringweather.base.Result
-import com.sunasterisk.boringweather.data.local.model.City
-import com.sunasterisk.boringweather.data.local.model.CurrentWeather
+import com.sunasterisk.boringweather.data.model.City
+import com.sunasterisk.boringweather.data.model.CurrentWeather
 
 interface CurrentRepository {
     fun getCurrentWeather(
