+++ app/src/main/java/com/sunasterisk/boringweather/data/repository/CurrentRepositoryImpl.kt
@@ -1,12 +1,13 @@
-package com.sunasterisk.boringweather.data
+package com.sunasterisk.boringweather.data.repository
 
 import android.os.AsyncTask
 import com.sunasterisk.boringweather.base.CallbackAsyncTask
 import com.sunasterisk.boringweather.base.Result
-import com.sunasterisk.boringweather.data.local.model.City
-import com.sunasterisk.boringweather.data.local.model.CurrentWeather
+import com.sunasterisk.boringweather.data.model.City
+import com.sunasterisk.boringweather.data.model.CurrentWeather
 
-class CurrentRepositoryImpl : CurrentRepository {
+class CurrentRepositoryImpl :
+    CurrentRepository {
 
     private var asyncTask: CallbackAsyncTask<*, *>? = null
