+++ app/src/main/java/com/sunasterisk/boringweather/data/repository/MainRepository.kt
@@ -1,0 +1,4 @@
+package com.sunasterisk.boringweather.data.repository
+
+interface MainRepository {
+}
