+++ app/src/main/java/com/sunasterisk/boringweather/ui/current/CurrentFragment.kt
@@ -9,12 +9,12 @@ import androidx.core.widget.NestedScrollView
 import com.google.android.material.navigation.NavigationView
 import com.sunasterisk.boringweather.R
 import com.sunasterisk.boringweather.base.BaseFragment
-import com.sunasterisk.boringweather.data.CityRepositoryImpl
-import com.sunasterisk.boringweather.data.CurrentRepositoryImpl
-import com.sunasterisk.boringweather.data.local.model.City
-import com.sunasterisk.boringweather.data.local.model.DailyWeather
-import com.sunasterisk.boringweather.data.local.model.HourlyWeather
-import com.sunasterisk.boringweather.data.local.model.SummaryWeather
+import com.sunasterisk.boringweather.data.repository.CityRepositoryImpl
+import com.sunasterisk.boringweather.data.repository.CurrentRepositoryImpl
+import com.sunasterisk.boringweather.data.model.City
+import com.sunasterisk.boringweather.data.model.DailyWeather
+import com.sunasterisk.boringweather.data.model.HourlyWeather
+import com.sunasterisk.boringweather.data.model.SummaryWeather
 import com.sunasterisk.boringweather.util.UnitSystem
 import com.sunasterisk.boringweather.util.showToast
 import com.sunasterisk.boringweather.util.verticalScrollProgress
